Solucionado el problema de los valores de linea.
Ahora se usan valores decimales

+++ ADOL.APP/ADOL.APP.BusinessEntities/Extensions/ApuestasDeportiva.cs
@@ -21,7 +21,7 @@ namespace ADOL.APP.CurrentAccountService.BusinessEntities
             }
         }
 
-        public float GetOddPrice(string oddType)
+        public decimal GetOddPrice(string oddType)
         {
             return this.OddProvider.GetOddValue(oddType, this);
         }
